Consolidates tidy-function files
